OdomF2M: added Odom/KeyFrameThr=0.5 and OdomF2M/MaxNewFeatures=0 parameters. Set Odom/GuessMotion to false by default.

+++ corelib/include/rtabmap/core/OdometryF2F.h
@@ -50,7 +50,7 @@ private:
 
 private:
 	//Parameters:
-	int keyFrameThr_;
+	float keyFrameThr_;
 
 	Registration * registrationPipeline_;
 	Signature refFrame_;
